requirements.txt
ffmpeg-python
matplotlib
numpy
pillow
pycuda
tensorflow[and-cuda]
tf_keras